Feature issue 31: Code cleanup with sonarqube

See merge request s142427/cloudguard!30

+++ api/src/main/java/com/cloudmen/cloudguard/dto/licenses/LicenseOverviewResponse.java
@@ -1,9 +1,9 @@
 package com.cloudmen.cloudguard.dto.licenses;
 
 public record LicenseOverviewResponse(
-        long totalAssigned,
-        long riskyAccounts,
-        long unusedLicenses,
-        long mfaPercentage
+        int totalAssigned,
+        int riskyAccounts,
+        int unusedLicenses,
+        int mfaPercentage
 ) {
 }
